backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
SQLAlchemy
pydantic
asyncio
